orchestration/requirements.txt
mlflow
scikit-learn
pandas
numpy
pyyaml
fastapi
uvicorn

